+++ pyVoIP/__init__.py
@@ -32,7 +32,15 @@ def debug(s, e=None):
 # noqa because import will fail if debug is not defined
 from pyVoIP.RTP import PayloadType  # noqa: E402
 
-SIPCompatibleMethods = ["INVITE", "ACK", "BYE", "CANCEL", "OPTIONS"]
+SIPCompatibleMethods = [
+    "INVITE",
+    "ACK",
+    "BYE",
+    "CANCEL",
+    "OPTIONS",
+    "SUBSCRIBE",
+    "NOTIFY",
+]
 SIPCompatibleVersions = ["SIP/2.0"]
 
 RTPCompatibleVersions = [2]