+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 1.7.2+RFC
 
+- SIP - More thorough subscription support
 - VoIP - Add DTMF support
 - VoIP - Backport authentication support from pyVoIP 2.x.x
 - VoIP - Hangup call support